fix a subtle bug concerning spurious timeouts where in_flight would go negative.
In a particular test, the sequence was this:

 - five packets suffer spurious retransmit timeouts.  This isn't by
   itself a problem - we handle this correctly.

 - the timed out packets can't be sent immediately due to lack of
   cwnd, so they get queued.

 - while two of them are still queued, they are cumulatively acked.

Now, when we added these packets to the retransmit queue, we reduced
_in_flight by pktsize because we decided they had left the network.

When the ack arrived, we correctly removed them from the retransmit
queue.  But we didn't undo the reduction in _in_flight, so recvd_bytes
decreased _in_flight for packets we'd already reduced it for.

The quick solution is to increase _in_flight by pktsize when an ack
covers a packet that is in the rtx queue.

In tracking this down, I realized we only search the rtx queue for
packets that are cumulatively acked, not ones that are selectively
acked, so I fixed that too.