Logfile::transposeLog: fix memory leak of 'val3Rec'  (#50)
* Logfile::transposeLog: fix memory leak of 'val3Rec'

When running the commit checks, Address Sanitizer reports:

  Direct leak of 16 byte(s) in 1 object(s) allocated from:
      #0 0x7ffff74b6357 in operator new[](unsigned long) ../../../../src/libsanitizer/asan/asan_new_delete.cpp:102
      #1 0x555555e4d972 in Logfile::transposeLog() htsim/sim/logfile.cpp:90
      #2 0x555555e4f70b in Logfile::~Logfile() htsim/sim/logfile.cpp:40
      #3 0x555555cfb2ae in main htsim/sim/datacenter/main_uec.cpp:552
      #4 0x7ffff6429d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58

While the other dynamically allocated locals were being deleted at the end of
this function, `val3Rec` was missed. Instead of adding in the manual deletion,
using RAII to avoid having to track this is an improvement.

Reported-by: GCC 11.4.0 ASan

* Logfile::transposeLog: remove manual memory management

This should prevent future occurrences of bugs like that fixed in the previous
commit.

* Logfile::transposeLog: remove a no-op pointer operation