fix 'parse_output' memory leaks (#69)
* parse_output: fix string memory leaks

Github: #56 “Work on cleanup up all memory issues as displayed by valdgrind”
Reported-by: GCC 11.4.0 ASan

* parse_output: fix memory leak of 'val3Rec'

This code and logfile.cpp seem to have stemmed from the same source, hence this
containing the same memory leak that was fixed in logfile.cpp in
22550d6d39556c95e74bb58a75b7e30f4751ab3d.

While the other dynamically allocated locals were being deleted at the end of
this function, `val3Rec` was missed. Instead of adding in the manual deletion,
using RAII to avoid having to track this is an improvement.

Github: #56 “Work on cleanup up all memory issues as displayed by valdgrind”
Reported-by: GCC 11.4.0 ASan

* parse_output: remove some manual memory management

This should prevent future occurrences of bugs like that fixed in the previous
commit.

This is the equivalent of changes applied in
22550d6d39556c95e74bb58a75b7e30f4751ab3d to logfile.cpp

Github: #56 “Work on cleanup up all memory issues as displayed by valdgrind”