remove 'UecNIC::_num_queued_srcs' that duplicates 'UecNIC::_active_srcs.size()' (#85)
This member is only adjusted when `_active_srcs` is pushed-to/popped-from. Its
value always mirrors the size of `_active_srcs`. Removing this manual size
tracking removes a possibility for bugs that cause the two to drift out of sync,
such as that fixed in 4b45235ff8b217132d6ca0b15e19b30f7f911935.

Commented out code is updated here as well, under the assumption this may be
un-commented in future.