UecSrc::multiplicate_decrease: fix missing '_cwnd' clamping
§1.1.13.6 “NSCC Internal Functions” in the latest Congestion Management
Specification (v0.9) gives the pseudocode of this function as:

  multiplicative_decrease():
    ccc.increase = FALSE # turn off fast increase
    ccc.fi_count = 0
    avg_delay = get_avg_delay()
    if avg_delay > target_qdelay:
      if (now - ccc.last_dec_time) > ccc.base_rtt:
        ccc.cwnd *= max(1-gamma*(avg_delay-
                  target_qdelay)/avg_delay,
                  max_md_jump)
        ccc.cwnd = max(ccc.cwnd, MTU)
        ccc.last_dec_time = now

In the implementation in htsim, the `ccc.cwnd = max(ccc.cwnd, MTU)` logic was
omitted from `multiplicative_decrease` and instead handled in the caller,
`updateCwndOnAck_NSCC`. This was functionally equivalent but harder to relate to
the specification and also missed updating the statistics counters.

Tracing the history of this in the specification, v0.5 did not have this
clamping, v0.6 was missing pseudocode for `multiplicative_decrease`, then it
reappeared in v0.75 with the clamping operation.